# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Skill {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  content: string;
  /** @nullable */
  inputSnippet?: string | null;
  /** @nullable */
  tokenCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface SkillInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  content: string;
  inputSnippet?: string;
  tokenCount?: number;
}

export interface SkillUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minLength 1 */
  content?: string;
  tokenCount?: number;
}

export interface SkillStats {
  totalSkills: number;
  recentCount: number;
}

export interface Template {
  id: number;
  name: string;
  description: string;
  content: string;
  category: string;
  createdAt: string;
}

export type ListTemplatesParams = {
category?: string;
};