# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  content: string;
  /** @nullable */
  inputSnippet?: string | null;
  /** @nullable */
  tokenCount?: number | null;
  createdAt: Date;
  updatedAt: Date;
}