# lib/api-zod/src/generated/types/skillInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  content: string;
  inputSnippet?: string;
  tokenCount?: number;
}