# lib/api-zod/src/generated/types/skillUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minLength 1 */
  content?: string;
  tokenCount?: number;
}