# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: number;
  name: string;
  description: string;
  content: string;
  category: string;
  createdAt: Date;
}